pyproject.toml
[build-system]
requires = ["hatchling>=1.18.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["py_ocpi"]
include = ["README.md", "LICENSE"]

[tool.hatch.build.targets.sdist]
exclude = [
  "/docs/",
  "/.github/",
  "/tests/",
  "/.gitignore",
  "/.pre-commit",
  "/Pipfile",
  "/Pipfile.lock",
  "/readthedocs.yaml",
]

[tool.black]
line-length = 80
target-version = ["py313"]
include = '\.pyi?$'
exclude = '''
(
  /(
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    |^temp|
    |ignore
  )/
  | pyproject.toml
)
'''

[project]
name = "extrawest_ocpi"
description = "Python implementation of Open Charge Point Interface (OCPI) protocol based on fastapi."
requires-python = ">=3.13"
readme = "README.md"
license = {text = "MIT"}
dynamic = ["version"]
dependencies = [
    "fastapi>=0.115.0",
    "httpx>=0.27.0",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
]

[[project.authors]]
name = "Oleksandr Bozbei"
email = "oleksandr.bozbei@extrawest.com"

[project.urls]
Homepage = "https://github.com/extrawest/extrawest_ocpi"
Repository = "https://github.com/extrawest/extrawest_ocpi"

[dependency-groups]
dev = [
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
]

[tool.hatch.version]
path = "py_ocpi/__init__.py"

[tool.flake8]
max-line-length = 88
per-file-ignores = [
    '__init__.py:F401',
]
count = true
exclude = "tests/*"

[tool.mypy]
python_version = "3.14"
ignore_missing_imports = true
plugins = ["pydantic.mypy"]
exclude = ["venv", "tests"]

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "DEBUG"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
testpaths = "tests"
addopts = [
    "--import-mode=importlib",
]
pythonpath = [
  "." ,"ocpi"
]
